Fix: Ahora los precios son con coma

+++ types/category.go
@@ -3,8 +3,8 @@ package types
 type Category struct {
 	Id       		string		`json:"id"`
 	Name     		string 		`json:"name"`
-	MaxAmount		int   		`json:"maxAmount"`
-	SpentAmount 	int 		`json:"spentAmount"`
+	MaxAmount		float32   		`json:"maxAmount"`
+	SpentAmount 	float32 		`json:"spentAmount"`
 }
 
 func (c Category) GetId() string{
